wip: move global navigation to core dir
